Move views to its own files
