+++ podcasts/New Search/Views/Results/SearchEpisodeCell.swift
@@ -1,0 +1,52 @@
+import SwiftUI
+import PocketCastsDataModel
+import PocketCastsUtils
+
+struct SearchEpisodeCell: View {
+    @EnvironmentObject var theme: Theme
+
+    var podcast: Podcast
+
+    var episode: Episode
+
+    var body: some View {
+        ZStack {
+            Button(action: {
+                print("row tapped")
+            }) {
+                Rectangle()
+                    .foregroundColor(.clear)
+                    .frame(maxWidth: .infinity, maxHeight: .infinity)
+            }
+            .buttonStyle(ListCellButtonStyle())
+
+            VStack(alignment: .leading, spacing: 12) {
+                HStack(spacing: 12) {
+                    PodcastCover(podcastUuid: podcast.uuid)
+                        .frame(width: 48, height: 48)
+                        .allowsHitTesting(false)
+                    VStack(alignment: .leading, spacing: 2) {
+                        Text(DateFormatHelper.sharedHelper.tinyLocalizedFormat(episode.publishedDate).localizedUppercase)
+                            .font(style: .footnote, weight: .bold)
+                            .foregroundColor(AppTheme.colorForStyle(.primaryText02, themeOverride: theme.activeTheme).color)
+                        Text(podcast.title ?? "")
+                            .font(style: .subheadline, weight: .medium)
+                            .foregroundColor(AppTheme.colorForStyle(.primaryText01, themeOverride: theme.activeTheme).color)
+                            .lineLimit(2)
+                        Text(TimeFormatter.shared.multipleUnitFormattedShortTime(time: TimeInterval(episode.duration)))
+                            .font(style: .caption, weight: .semibold)
+                            .foregroundColor(AppTheme.colorForStyle(.primaryText02, themeOverride: theme.activeTheme).color)
+                            .lineLimit(1)
+                    }
+                    .allowsHitTesting(false)
+                }
+                .padding(.trailing, 16)
+                ThemeableSeparatorView()
+            }
+            .padding(EdgeInsets(top: 12, leading: 16, bottom: 0, trailing: 0))
+            .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0))
+            .listRowSeparator(.hidden)
+            .listSectionSeparator(.hidden)
+        }
+    }
+}
